refactor: WIP move version download code to use point
